ScitePipeline/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
supabase>=2.3.0
httpx>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
transformers>=4.36.0
torch>=2.1.0
google-genai>=1.0.0

